Two things: most of Prerex/Link checking; added continued blocks.
Prerex/Link checking is in Check.hs.  It's the first layer of static
checking for idoc, the other only being able to be done once the
database is online.

Continued Blocks is a modification to some existing blocks to allow
optional (and sometimes manditory) "continuations" of those blocks.
This is useful, for example, when you wish to state a theorem and then
immediately prove it.