Switched to Vinyl for Blocks.
Big changes in this branch, though it's not finished yet and doesn't
compile/run.

Vinyl `CoRec's are used now instead of builtin ADTs for blocks.  This
makes blocks "pluggable" in the sense that more can be added in a
library, and a new language that lacks some can be defined in a
library as well.  Also, their textual representation is completely
open to the user of the library, though the old default exists as a
funciton in `Parse.hs', and will be used for ILS-flavoured idoc.

Using some type magic in `Syntax.hs', I was able to ensure that it was
possible to render any combination of blocks in a `CoRec', so long as
each blocktype in the `CoRec' implements the new typeclass
`BlockMarkup' (speaking only of Html5; Tex is completely broken at
this point).

Implemented this for a lot of blocks, but still need to finish up,
starting at `Blocks/Math.hs' (alphabetically).